refactor : constant 패키지 하위로 Message enum 옮김
