+++ src/main/java/app/bottlenote/review/dto/response/ReviewReplyResponse.java
@@ -1,5 +1,6 @@
 package app.bottlenote.review.dto.response;
 
+import app.bottlenote.review.dto.response.constant.ReviewReplyResultMessage;
 import java.time.LocalDateTime;
 import java.time.format.DateTimeFormatter;
